Temporary merge all components into orders page
